# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BidResult": {
        "properties": {
          "cost": {
            "additionalProperties": true,
            "title": "Cost",
            "type": "object"
          },
          "fields": {
            "additionalProperties": {
              "additionalProperties": {
                "$ref": "#/components/schemas/FieldValue"
              },
              "type": "object"
            },
            "title": "Fields",
            "type": "object"
          },
          "review_confirmed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Confirmed By"
          },
          "ruleset_version": {
            "title": "Ruleset Version",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "stage1": {
            "$ref": "#/components/schemas/StageSummary"
          },
          "stage2": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "tenderer": {
            "title": "Tenderer",
            "type": "string"
          },
          "trace": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace"
          },
          "verdicts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Verdict"
            },
            "title": "Verdicts",
            "type": "object"
          }
        },
        "required": [
          "tenderer",
          "run_id",
          "ruleset_version",
          "fields",
          "verdicts",
          "stage1"
        ],
        "title": "BidResult",
        "type": "object"
      },
...
      "CheckRequest": {
        "properties": {
          "tenderers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "default: every tenderer with an uploaded offer",
            "title": "Tenderers"
          }
        },
        "title": "CheckRequest",
        "type": "object"
      },
      "CheckResponse": {
        "properties": {
          "job_ids": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Job Ids",
            "type": "object"
          }
        },
        "required": [
          "job_ids"
        ],
        "title": "CheckResponse",
        "type": "object"
      },
...
      "CheckedField": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "the field's key inside BidResult.fields[letter]",
            "title": "Field"
          },
          "field_id": {
            "title": "Field Id",
            "type": "string"
          },
          "follow_up": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "redacted": {
            "default": false,
            "title": "Redacted",
            "type": "boolean"
          },
          "stage": {
            "default": "I",
            "title": "Stage",
            "type": "string"
          },
          "status": {
            "enum": [
              "pass",
              "needs_review",
              "disqualified",
              "dormant"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "field_id",
          "status"
        ],
        "title": "CheckedField",
        "type": "object"
      },
...
      "Correction": {
        "properties": {
          "by": {
            "title": "By",
            "type": "string"
          },
          "model_value": {
            "title": "Model Value"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "by",
          "reason"
        ],
        "title": "Correction",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Every error response, 422 included (docs/api_contract.md, Conventions).",
        "properties": {
          "detail": {
            "description": "the message again, for the Streamlit UI until S5",
            "title": "Detail",
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error",
          "detail"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "snake_case, e.g. validation_failed",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "FieldValue": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "correction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Correction"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_value": {
            "title": "Model Value"
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageCitation"
              },
              {
                "type": "null"
              }
            ]
          },
          "redacted": {
            "default": false,
            "title": "Redacted",
            "type": "boolean"
          },
          "value": {
            "title": "Value"
          }
        },
        "title": "FieldValue",
        "type": "object"
      },
...
      "Job": {
        "properties": {
          "attempt": {
            "default": 1,
            "title": "Attempt",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "progress": {
            "$ref": "#/components/schemas/Progress"
          },
          "project": {
            "title": "Project",
            "type": "string"
          },
          "ruleset_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ruleset Version"
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "paused",
              "done",
              "failed",
              "dead"
            ],
            "title": "State",
            "type": "string"
          },
          "step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          },
          "tenderer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenderer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "kind",
          "project",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "Job",
        "type": "object"
      },
      "JobPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "JobPage",
        "type": "object"
      },
...
      "Progress": {
        "description": "How far a job's current step has got: `done` of `total` `unit`s (pages, calls, ...).",
        "properties": {
          "done": {
            "default": 0,
            "title": "Done",
            "type": "integer"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "title": "Progress",
        "type": "object"
      },
...
      "StageSummary": {
        "properties": {
          "items": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Items",
            "type": "object"
          },
          "outcome": {
            "enum": [
              "pass",
              "needs_review",
              "disqualified",
              "dormant"
            ],
            "title": "Outcome",
            "type": "string"
          }
        },
        "required": [
          "outcome",
          "items"
        ],
        "title": "StageSummary",
        "type": "object"
      },
...
      "Verdict": {
        "properties": {
          "checks": {
            "items": {
              "$ref": "#/components/schemas/CheckedField"
            },
            "title": "Checks",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/PageCitation"
            },
            "title": "Evidence",
            "type": "array"
          },
          "outcome": {
            "enum": [
              "pass",
              "needs_review",
              "disqualified",
              "dormant"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "part": {
            "title": "Part",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "rule_ids": {
            "items": {
              "type": "string"
            },
            "title": "Rule Ids",
            "type": "array"
          },
          "worst": {
            "enum": [
              "pass",
              "needs_review",
              "disqualified",
              "dormant"
            ],
            "title": "Worst",
            "type": "string"
          }
        },
        "required": [
          "outcome",
          "worst",
          "part",
          "rule_ids",
          "reason",
          "checks",
          "evidence"
        ],
        "title": "Verdict",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Tender Evaluation Assistant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/projects/{pid}/bids/{tenderer}/results": {
      "get": {
        "operationId": "bid_results_projects__pid__bids__tenderer__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenderer",
            "required": true,
            "schema": {
              "title": "Tenderer",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BidResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bid Results"
      }
    },
...
    "/projects/{pid}/checks": {
      "post": {
        "description": "One job per tenderer against the latest confirmed rule set (open question 3:\nlatest confirmed, `?version=` only on evaluate).",
        "operationId": "start_checks_projects__pid__checks_post",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "x-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CheckRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Checks"
      }
    },
...
    "/projects/{pid}/evaluation": {
      "get": {
        "operationId": "get_evaluation_projects__pid__evaluation_get",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Evaluation Projects  Pid  Evaluation Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evaluation"
      }
    },
...
    "/projects/{pid}/jobs": {
      "get": {
        "operationId": "list_jobs_projects__pid__jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs"
      }
    },
    "/projects/{pid}/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_projects__pid__jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
...
    "/projects/{pid}/legacy/evaluate": {
      "post": {
        "operationId": "run_evaluation_projects__pid__legacy_evaluate_post",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Evaluation Projects  Pid  Legacy Evaluate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Evaluation"
      }
    },
...
    "/projects/{pid}/usage": {
      "get": {
        "description": "Model usage and cost for the last run: totals, $ per bid, per-bid and per-model\nbreakdown, failed calls (fallbacks) — written by the graph next to evaluation.json.",
        "operationId": "get_usage_projects__pid__usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Usage Projects  Pid  Usage Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Usage"
      }
    }
  }
}